server/requirements.txt
Pillow
bleach==3.1.0
build-utils==0.3.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
docutils==0.14
fake-rpi==0.6.1
idna==2.8
Inject==3.5.1
itsdangerous==1.1.0
nose==1.3.7
pigpio==1.42
pkginfo==1.5.0.1
Pygments==2.4.2
requests==2.22.0
requests-toolbelt==0.9.1
RPi.GPIO==0.6.5
six==1.12.0
tqdm==4.32.1
twine==1.13.0
webencodings==0.5.1
websockets==7.0
